This part of the work is to help users locate the nearest police station incase they want to report an accident. Usersa can quickly serch for their location by entering the state where they are and can quickly locate the local government area, once the user enters the local government area, they can now easily check for the closest police station to report the accident
This also goes for Federal Road Safety Commission, as users can easly
get the contact number of frsc and inform them of any road accident they
just witnessed.

Users can also get the closest hospital incase of an emergency for quick
and rapid response in helping to save the life of an accident victim,
and in any situation of a loss of life, mortuary services are also
indicated.

In view of a user having to spend the nght in any town or city, they can
easily identify hotels in that place by checking for the hotel closest
to where they are.'#' will be ignored, and an empty message aborts the commit.